chatbot-server/requirements.txt
#Basic
python-dotenv
types-requests==2.31.0.20240406

# OpenAI
openai

# LangChain
nltk
langchain
langchain-openai

# Azure
azure-search-documents==11.6.0b4
azure-storage-blob
azure-identity
azure-keyvault-secrets

# Web-app
Flask==3.0.3
Flask-Cors==4.0.1

# Document Structuring
pdf2image